test(repo-graph): split source-read cap tests into a dedicated file
The FR-006 line-cap ratchet forbids over-cap test files from growing
(repo-graph-query.test.ts: 545 lines). Move the issue #2399 cap
coverage (oversized skip, at-cap boundary, oversized neighbor in a
mixed pack) into tests/unit/tools/repo-graph-query-source-cap.test.ts.
